# api/openapi.yaml
openapi: 3.1.0
info:
  title: ZATRANO API
  version: "1.0.0"
  description: |
    Clean/Hexagonal mimariyle, CQRS ve DDD prensipleriyle inşa edildi. Tüm hatalar RFC 7807 (Problem Details) formatında döner.
    Başarı yanıtları ve e-posta şablonları Accept-Language / ?lang ile yerelleştirilir.

    **Idempotency:** `POST`/`PUT`/`PATCH`/`DELETE` isteklerinde `Idempotency-Key` başlığı gönderilirse
    aynı anahtarla yapılan tekrar istekler ilk yanıtın aynısını döner (çift işlem engeli).
    Web karşılığı: GoUI session/component event koruması ve aynı application use-case idempotency kurallarıdır.
  license:
    name: MIT
servers:
  - url: http://localhost:8080
    description: Yerel geliştirme
...
paths:
...
  /api/v1/contact:
    post:
      tags: [public]
      summary: İletişim formu gönderimi
      description: |
        Mesajı veritabanına kaydeder ve CONTACT_RECIPIENT_EMAIL adresine
        kalıcı outbox üzerinden e-posta kuyruğuna alır. Turnstile (yapılandırılmışsa) zorunludur.
      parameters:
        - in: header
          name: X-Turnstile-Token
          schema: { type: string }
          description: Cloudflare Turnstile yanıt token'ı (opsiyonel body alanı da kabul edilir)
        - in: header
          name: Idempotency-Key
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, email, message]
              properties:
                name: { type: string, minLength: 2, maxLength: 100 }
                email: { type: string, format: email }
                message: { type: string, minLength: 5, maxLength: 2000 }
                locale: { type: string, example: tr }
      responses:
        "201":
          description: Mesaj alındı
          content:
            application/json:
              schema:
                type: object
                properties:
                  id: { type: string, format: uuid }
                  message: { type: string }
        "400": { $ref: "#/components/responses/Problem" }
        "429": { $ref: "#/components/responses/Problem" }
...
  /api/v1/auth/login:
    post:
      tags: [auth]
      summary: Giriş yap
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Token çifti
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TokenPair" }
        "401":
          description: Geçersiz kimlik bilgileri veya MFA challenge
          content:
            application/json:
              schema:
                oneOf:
                  - { $ref: "#/components/schemas/Problem" }
                  - { $ref: "#/components/schemas/MFARequiredResponse" }
        "429": { $ref: "#/components/responses/Problem" }
...
  /api/v1/auth/forgot-password:
    post:
      tags: [auth]
      summary: Şifre sıfırlama bağlantısı iste (numaralandırmaya karşı jenerik yanıt)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ForgotPasswordRequest" }
      responses:
        "200":
          description: Yerelleştirilmiş başarı mesajı
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MessageResponse" }
  /api/v1/auth/reset-password:
    post:
      tags: [auth]
      summary: Token ile yeni şifre belirle
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ResetPasswordRequest" }
      responses:
        "200":
          description: Yerelleştirilmiş başarı mesajı
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MessageResponse" }
        "401": { $ref: "#/components/responses/Problem" }
  /api/v1/auth/verify-email:
    post:
      tags: [auth]
      summary: E-posta adresini doğrula
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/TokenRequest" }
      responses:
        "200":
          description: Yerelleştirilmiş başarı mesajı
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MessageResponse" }
        "401": { $ref: "#/components/responses/Problem" }
...
  /api/v1/auth/oauth/{provider}:
    get:
      tags: [auth]
      summary: OAuth/SSO akışını başlat (sağlayıcıya yönlendirir)
      parameters:
        - { name: provider, in: path, required: true, schema: { type: string, enum: [google, github] } }
      responses:
        "302": { description: Sağlayıcının yetkilendirme sayfasına yönlendirme }
        "422": { $ref: "#/components/responses/Problem" }
  /api/v1/auth/oauth/{provider}/callback:
    get:
      tags: [auth]
      summary: OAuth/SSO callback (kod → token çifti)
      parameters:
        - { name: provider, in: path, required: true, schema: { type: string, enum: [google, github] } }
        - { name: code, in: query, required: true, schema: { type: string } }
        - { name: state, in: query, required: true, schema: { type: string } }
      responses:
        "200":
          description: Token çifti
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TokenPair" }
        "401": { $ref: "#/components/responses/Problem" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    Problem:
      description: RFC 7807 hata yanıtı
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
  schemas:
    Problem:
      type: object
      properties:
        type: { type: string }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string }
        code: { type: string }
        errors:
          type: array
          items:
            type: object
            properties:
              field: { type: string }
              message: { type: string }
...
    Email:
      type: string
      format: email
      maxLength: 254
      description: Normalize edilmiş küçük harf e-posta (pkg/validation.NormalizeEmail)
...
    TurnstileToken:
      type: string
      description: Cloudflare Turnstile token (cf-turnstile-response). SEC_TURNSTILE yapılandırıldığında zorunlu.
    LoginRequest:
      type: object
      description: İlk adımda email+password; MFA adımında mfa_challenge+mfa_code gönderilir.
      properties:
        email: { $ref: "#/components/schemas/Email" }
        password: { type: string }
        mfa_code:
          type: string
          description: 6 haneli TOTP veya tek kullanımlık kurtarma kodu
        mfa_challenge:
          type: string
          description: İlk adımın 401 yanıtından alınan, 5 dakika geçerli opak challenge
        turnstile_token: { $ref: "#/components/schemas/TurnstileToken" }
      oneOf:
        - required: [email, password]
        - required: [mfa_challenge, mfa_code]
    MFARequiredResponse:
      type: object
      required: [type, status, detail, mfa_challenge]
      properties:
        type: { type: string, example: auth.mfa_required }
        title: { type: string }
        status: { type: integer, example: 401 }
        detail: { type: string }
        mfa_challenge: { type: string }
    ForgotPasswordRequest:
      type: object
      required: [email]
      properties:
        email: { $ref: "#/components/schemas/Email" }
        turnstile_token: { $ref: "#/components/schemas/TurnstileToken" }
    ResetPasswordRequest:
      type: object
      required: [token, new_password]
      properties:
        token: { type: string }
        new_password: { type: string, minLength: 8, maxLength: 128 }
        turnstile_token: { $ref: "#/components/schemas/TurnstileToken" }
...
    TokenRequest:
      type: object
      required: [token]
      properties:
        token: { type: string }
...
    TokenPair:
      type: object
      properties:
        access_token: { type: string }
        refresh_token: { type: string }
        expires_at: { type: string, format: date-time }
        token_type: { type: string, example: Bearer }
...
    MessageResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string
          description: Accept-Language / ?lang ile yerelleştirilmiş başarı mesajı
...